rename component, styledtablecontainer
