+++ apps/portals/standards/src/pages/StandardsDetailsPage.tsx
@@ -6,7 +6,7 @@ import { DetailsPageContextConsumer } from '@sage-bionetworks/synapse-portal-fra
 import DetailsPage from '@sage-bionetworks/synapse-portal-framework/components/DetailsPage/index'
 import { useGetPortalComponentSearchParams } from '@sage-bionetworks/synapse-portal-framework/utils/UseGetPortalComponentSearchParams'
 import {
-  CardMetadataDisplay,
+  RowDataTable,
   ErrorPage,
   GenericCardSchema,
   SynapseConstants,
@@ -38,7 +38,7 @@ export const standardDetailsPageContent: DetailsPageContentType = [
     element: (
       <DetailsPageContextConsumer columnName={'Name'}>
         {({ value }) => (
-          <CardMetadataDisplay
+          <RowDataTable
             sql={dataSql}
             labels={[
               'Name',
