--- packages/synapse-react-client/src/components/CardMetadataDisplay/index.ts
@@ -1,4 +1,0 @@
-import CardMetadataDisplay from './CardMetadataDisplay'
-import type { CardMetadataDisplayProps } from './CardMetadataDisplay'
-export { CardMetadataDisplay, CardMetadataDisplayProps }
-export default CardMetadataDisplay
