+++ packages/synapse-react-client/src/components/RowDataTable/RowDataTable.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta, StoryObj } from '@storybook/react'
-import CardMetadataDisplay from './CardMetadataDisplay'
+import RowDataTable from './RowDataTable'
 
 const dataColumnAliases = {
   name: 'Name',
@@ -8,8 +8,8 @@ const dataColumnAliases = {
 }
 
 const meta = {
-  title: 'UI/CardMetadataDisplay',
-  component: CardMetadataDisplay,
+  title: 'UI/RowDataTable',
+  component: RowDataTable,
   parameters: {
     chromatic: { viewports: [600, 1200] },
   },
@@ -19,7 +19,7 @@ export default meta
 type Story = StoryObj<typeof meta>
 
 export const Demo: Story = {
-  render: args => <CardMetadataDisplay {...args} />,
+  render: args => <RowDataTable {...args} />,
   args: {
     sql: 'SELECT * FROM syn52694652',
     labels: ['name', 'dataType', 'grantDOI'],
