+++ packages/synapse-react-client/src/components/RowDataTable/index.ts
@@ -1,0 +1,4 @@
+import RowDataTable from './RowDataTable'
+import type { RowDataTableProps } from './RowDataTable'
+export { RowDataTable, RowDataTableProps }
+export default RowDataTable
