+++ packages/synapse-react-client/src/components/index.ts
@@ -5,7 +5,7 @@ export * from './AccountLevelBadges'
 export * from './Authentication'
 export * from './CardContainer'
 export * from './CardContainerLogic'
-export * from './CardMetadataDisplay'
+export * from './RowDataTable'
 export * from './CertificationQuiz'
 export * from './ChallengeDetailPage'
 export * from './ChallengeDataDownload'
